Separate typelib from typesys. Upgrade SemId system.
